// src/types/notification.ts
export interface NotificationData {
  title: string;
  description: string;
  link: string | null;
  type: "info" | "warning" | "success" | "error";
}

export interface Notification {
  id: string;
  type: string;
  notifiable_type: string;
  notifiable_id: number;
  data: NotificationData;
  read_at: string | null;
  created_at: string;
  updated_at: string;
}